requirements.txt
azure-storage-blob
pillow
python-dotenv
